Add @vercel/analytics package and enable web analytics
