+++ api/latest/index.mjs
@@ -1,0 +1,10 @@
+import fs from 'fs';
+
+export default async (request, response) => {
+    const updateData = fs.readFileSync('./src/update.json');
+    const updateJson = JSON.parse(updateData);
+
+    const highestId = Math.max(...updateJson.map(item => item.id));
+
+    response.status(200).send(JSON.stringify({ highestId }));
+};
