+++ package.json
@@ -32,6 +32,7 @@
     "@mui/material": "^5.14.17",
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
+    "@vercel/analytics": "^1.1.1",
     "@vercel/node": "^3.0.11",
     "astro": "^3.3.0",
     "express": "^4.18.2",
