+++ public/theme.css
@@ -1,13 +1,52 @@
 @import url('https://fonts.googleapis.com/css?family=Ubuntu:300,400,500,700');
 
-body {
-    font-family: 'Ubuntu', sans-serif;
-    font-size: 14px;
-    line-height: 1.42857143;
-    color: #3c108d;
-    background-image: "logo.png";
+:root {
+    --background: #877895;
+  }
+  body 
+  {
+    background: #877895;
+    font-family: Arial, sans-serif;
+    margin: 0;
+    padding: 0;
+    
+  }
+  a {
+    color: #c90000;
+    text-decoration: none;
+  }
+  a:hover {
+    color: #9b59b6;
+    text-decoration: underline;
+  }
+  h1, h2, h3, h4, h5, h6 {
+    color: #000000;
+    font-weight: bold;
+    font-family: Arial, sans-serif;
+  }
+  p {
+    color: #bdc3c7;
+    font-size: 16px;
+    line-height: 1.5;
+  }
+  .container {
+    max-width: 800px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  * {
+    box-sizing: border-box;
+    --background: #fa;
+  
+  }
+  
+  .notes {
+    margin: 0 auto;
+    max-width: 800px;
+    padding: 0 1rem;
+    background : var(--background);
+  }
 
-}
 .documentation .hero {
 
     position: relative;
