--- src/pages/api-reference.astro
@@ -1,106 +1,0 @@
----
-import Layout from '../layouts/Layout.astro';
----
-<link rel="stylesheet" href="theme.css">
-<!DOCTYPE html>
-<html lang="en">
-<head>
-   
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta property="og:title" content="CubeScript">
-    <meta property="og:description" content="the official API refrence infomation for CubeScript">
-    <meta property="og:image" content="https://cubescript.vercel.app/logo.png">
-    <meta property="og:image:width" content="1200">
-    <meta property="og:image:height" content="1080">
-    <meta property="og:url" content="https://CubeScript.vercel.app">
-    <title>getting-started</title>
-</head>
-<body>
-    
-    <div class="cat">
-
-        <Layout title="Getting Started with CubeScript">
-
-            <h4 style="color: white ;">Getting Started with CubeScript's ApiReference</h4><br>
-            
-   
-            <p>CubeScript Only has one library at the moment. hopefully more will be added soon.</p>
-            <br>
-            <p> CubeScript's Stdlib, comonly refered to as the "Standard Library" is a library that comes with CubeScript by default. it contains all the basic functions and commands that you will need to make a CubeScript script. </p>
-            <p>the Home.dll that comes with the Compiler is the Stdlib in question, and so because of that, we will be going over the essential documentation for the Stdlib</p>
-            <br>
-            <p>Within Home.dll contains 2 classes, Home and Letterbox.</p>
-            <p>Home doesn't contain much at the moment, but it does contain some basic maths and print/error functions</p>
-            <p>you can visit the documentation for Home.dll <a href="/docs/home">here</a></p>
-            <br>
-            <p>Letterbox is a class that contains all the functions and commands that you will need to make a CubeScript script. it contains functions for things like, printing, getting the input from a user, reading and writing to files and even creating and removing directorys.</p>
-        </div>
-
-
-
-
-
-            
-            
-            
-            
-            
-            
-            
-        </Layout>
-    </div>
-        <div class="bg-animation">
-        <div id="stars"></div>
-        <div id="stars2"></div>
-        <div id="stars3"></div>
-        <div id="stars4"></div>
-    </div><div class="bg-animation">
-        <div id="stars"></div>
-        <div id="stars2"></div>
-        <div id="stars3"></div>
-        <div id="stars4"></div>
-    </div>
-<style>
-
-    .windows {
-  background-color: #868686;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  margin: 20px auto;
-  max-width: 200px;
-  padding: 20px;
-
-}
-.windows a{
-    color: #fff;
-    background-color: #007bff;
-    border-color: #007bff;
-    display: inline-block;
-    font-weight: 400;
-    text-align: center;
-    vertical-align: center;
-    cursor: pointer;
-    user-select: none;
-    border: 1px solid transparent;
-    padding: .375rem .75rem;
-    font-size: 1rem;
-    line-height: 1.5;
-    border-radius: .25rem;
-
-}
-    .cat
-    {
-        background-color: rgba(0, 255, 221, 0.438);
-        color: white;
-        text-align: center;
-        font-family: 'Courier New', Courier, monospace;
-        font-size: 20px;
-        padding: 20px;
-        border-radius: 10px;
-        margin: 20px;
-    }
-</style>
-</body>
-</html>
