+++ src/pages/index.html
@@ -9,6 +9,12 @@
 <head>
 	<title>Dimension by HTML5 UP</title>
 	<meta charset="utf-8" />
+	<meta property="og:title" content="CubeScript">
+	<meta property="og:description" content="the official Documetation infomation for CubeScript">
+	<meta property="og:image" content="https://cubescript.vercel.app/logo.png">
+	<meta property="og:image:width" content="1200">
+	<meta property="og:image:height" content="1080">
+	<meta property="og:url" content="https://CubeScript.vercel.app">
 	<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
 	<link rel="stylesheet" href="assets/css/main.css" />
 	<noscript>
@@ -36,11 +42,13 @@
 				<ul>
 					<li><a href="#intro">Home</a></li>
 					<li><a href="#work">CubeScript</a></li>
+					<li><a href="/tutorial/api-reference">API</a></li>
 					<li><a href="#about">About</a></li>
-					<li><a href="#Scoop">Scoop.sh</a></li>
+					<li><a href="Patchnotes">patchnotes</a></li>
+					<li><a href="Download">Download</a></li>
+					<!--
 					<li><a href="#contact">Contact</a></li>
-					<li><a href="#contact">Contact</a></li>
-					<li><a href="#contact">Contact</a></li>
+					<li><a href="#contact">Contact</a></li> -->
 				</ul>
 			</nav>
 		</header>
@@ -52,7 +60,11 @@
 			<article id="intro">
 				<h2 class="major">Intro</h2>
 				<span class="image main"></span>
-				<p>AWelcome to CubeScript!</p>
+				<p>AWelcome to the official CubeScript website! </p>
+				<p>CubeScript is a transpiled Programming language designed by OpenStudio</p>
+				<p>using the Probided links on the homepage, you are able to move around the site!</p>
+				<p>Currently this site is being worked on, but you can still visit the old pages by clicking on the buttons on the home page</p>
+				<p>slowly overtime, your most visited pages should look different. thats because we are overhauling the website</p>
 			</article>
 
 			<!-- Work -->
@@ -60,18 +72,16 @@
 				<h2 class="major">Work</h2>
 				<span class="image main"></span>
 				<p>CubeScript is a scripting language based around the C style family of languages.</p>
-				<p></p>
+				<p>CubeScript is known for having syntax like python, lua, CSharp and even typescript</p>
+				<p>the language is designed to be easy to learn, and easy to use</p>
 			</article>
 
 			<!-- About -->
 			<article id="about">
 				<h2 class="major">About</h2>
 				<span class="image main"></span>
-				<p>Lorem ipsum dolor sit amet, consectetur et adipiscing elit. Praesent eleifend dignissim arcu, at
-					eleifend sapien imperdiet ac. Aliquam erat volutpat. Praesent urna nisi, fringila lorem et vehicula
-					lacinia quam. Integer sollicitudin mauris nec lorem luctus ultrices. Aliquam libero et malesuada
-					fames ac ante ipsum primis in faucibus. Cras viverra ligula sit amet ex mollis mattis lorem ipsum
-					dolor sit amet.</p>
+				<p>Thanks to all the maintainers and developers at OpenStudio and OpenStudioCorporation for helping Create and Maintain the offical CubeScript website and the CSVM transpiler!</p>
+				<p>you can visit the official OpenStudio Website at <a href="https://OpenStudioCorp.vercel.app"></a></p>
 			</article>
